1. Changed workflow for creation and opening of encrypted databases.
When opening an encrypted database, the user is prompted to enter their original password rather than an encryption key/hash. The password is used to generate an encryption key and the database is opened using that key. If they happen to enter a Base64 encoded encryption key or a legacy encryption hash instead, the code catches that and takes the appropriate action (opens the db using the encryption key, or prompts them to upgrade the encryption of their db if it's a hash).

When creating an encrypted database, the dialog now indicates that the password will be used to open the database. After the db is created the popup still appears displaying the encryption key (Base64 encoded). However, the popup indicates that they encryption key is to use "to open the database in other applications, not in Lita".

2. Added public function getBase64FromPassword() to MainModel. This method can be used to "recover" an encryption key, given the original password. In fact it just performs the same steps that are used to create the encryption key using the SimpleEncryptionKeyGenerator class.